+++ CHANGES.md
@@ -14,6 +14,8 @@ To be released.
     `InvalidMessageSignatureException` and `InvalidMessageTimestampException`
     instead of `InvalidMessageException` and `InvalidTimestampException`
     respectively.  [[#1771]]
+ -  (Libplanet.Net) Added `long tipDeltaThreshold = 25L` option to
+    `Swarm<T>.PreloadAsync()` method.  [[#1775], [#1777], [#1779]]
 
 ### Backward-incompatible network protocol changes
 
@@ -29,7 +31,20 @@ To be released.
 
 ### CLI tools
 
-[#1771] https://github.com/planetarium/libplanet/pull/1771
+[#1771]: https://github.com/planetarium/libplanet/pull/1771
+[#1779]: https://github.com/planetarium/libplanet/pull/1779
+
+
+Version 0.27.1
+--------------
+
+Released on February 7, 2022.
+
+ -  (Libplanet.Net) `Swarm<T>.PreloadAsync()` method became to catch up recent
+    blocks until the local tip is close enough to the topmost tip among peers.
+    [[#1775], [#1777]]
+
+[#1777]: https://github.com/planetarium/libplanet/pull/1777
 
 
 Version 0.27.0
@@ -116,6 +131,18 @@ Released on January 26, 2022.
 [#1766]: https://github.com/planetarium/libplanet/pull/1766
 [CVE-2022-0235]: https://github.com/advisories/GHSA-r683-j2x4-v87g
 [Libplanet.Net]: https://www.nuget.org/packages/Libplanet.Net/
+
+
+Version 0.26.3
+--------------
+
+Released on February 7, 2022.
+
+ -  `Swarm<T>.PreloadAsync()` method became to catch up recent blocks
+    until the local tip is close enough to the topmost tip among peers.
+    [[#1775]]
+
+[#1775]: https://github.com/planetarium/libplanet/pull/1775
 
 
 Version 0.26.2